requirements.txt
python-binance==1.0.17
python-telegram-bot==13.15
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
APScheduler==3.10.1
